add c-stl

+++ stl_random.h
@@ -13,7 +13,7 @@ public:
 	static int random_int(int min, int max);
 	static double random_double(double min, double max);
 	static int frand_between(int min, int max, int step);
-	static double frand_between_double(int min, int max, int step);
+	static double frand_between_double(double min, double max, double step);
 private:
 	static mt19937 mt;
 };
